Pool searchMatchingEquipments
