+++ src/components/dialogs/equipment-search-dialog.js
@@ -10,20 +10,12 @@ import { useIntl } from 'react-intl';
 import {
     ElementSearchDialog,
     equipmentStyles,
-    getEquipmentsInfosForSearchBar,
     EquipmentItem,
 } from '@gridsuite/commons-ui';
-import React, { useCallback, useState, useRef } from 'react';
-import { fetchEquipmentsInfos } from '../../utils/rest-api';
-import {
-    displayErrorMessageWithSnackbar,
-    useIntlRef,
-} from '../../utils/messages';
+import React, { useRef } from 'react';
 import { useParams } from 'react-router-dom';
-import { PARAM_USE_NAME } from '../../utils/config-params';
 import makeStyles from '@mui/styles/makeStyles';
-import { useSnackbar } from 'notistack';
-import { useSelector } from 'react-redux';
+import { useSearchMatchingEquipments } from '../util/search-matching-equipments';
 
 const useEquipmentStyles = makeStyles(equipmentStyles);
 
@@ -45,51 +37,16 @@ const EquipmentSearchDialog = ({
     const equipmentClasses = useEquipmentStyles();
 
     const intl = useIntl();
-    const intlRef = useIntlRef();
-    const { enqueueSnackbar } = useSnackbar();
     const studyUuid = decodeURIComponent(useParams().studyUuid);
-    const useNameLocal = useSelector((state) => state[PARAM_USE_NAME]);
-    const [equipmentsFound, setEquipmentsFound] = useState([]);
     const lastSearchTermRef = useRef('');
-
-    const searchMatchingEquipments = useCallback(
-        (searchTerm) => {
-            lastSearchTermRef.current = searchTerm;
-            fetchEquipmentsInfos(
-                studyUuid,
-                currentNodeUuid,
-                searchTerm,
-                useNameLocal,
-                true,
-                equipmentType
-            )
-                .then((infos) => {
-                    if (searchTerm === lastSearchTermRef.current) {
-                        setEquipmentsFound(
-                            getEquipmentsInfosForSearchBar(infos, useNameLocal)
-                        );
-                    } // else ignore results of outdated fetch
-                })
-                .catch((errorMessage) =>
-                    displayErrorMessageWithSnackbar({
-                        errorMessage: errorMessage,
-                        enqueueSnackbar: enqueueSnackbar,
-                        headerMessage: {
-                            headerMessageId: 'equipmentsSearchingError',
-                            intlRef: intlRef,
-                        },
-                    })
-                );
-        },
-        [
+    const [searchMatchingEquipments, equipmentsFound] =
+        useSearchMatchingEquipments(
             studyUuid,
             currentNodeUuid,
-            useNameLocal,
-            enqueueSnackbar,
-            intlRef,
-            equipmentType,
-        ]
-    );
+            lastSearchTermRef,
+            true,
+            equipmentType
+        );
 
     return (
         <ElementSearchDialog
